Add mentoring combo button to <Nudge />
